tests/wpt/mozilla/tests/bluetooth/writeValue/descriptor/write-updates-value.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(generic_access.name))
    .then(service => service.getCharacteristic(device_name.name))
    .then(characteristic => characteristic.getDescriptor(number_of_digitals.name))
    .then(descriptor => {
        assert_equals(descriptor.value, null);
        return descriptor.writeValue(Uint8Array.of(1, 2))
        .then(() => assert_equals(descriptor.value, '\x01\x02'));
    });
}, 'A regular write request to a writable descriptor should update value.');
</script>
